Keep feature (#2)
* add keep feature

* test 2

* remove origin

* this is the one

* ReadME update

* typo

+++ entrypoint.sh
@@ -11,15 +11,17 @@ OUT=""
 for i in $(git for-each-ref refs/remotes/origin --sort=committerdate  --format='%(HEAD)%(color:yellow)%(refname:short)%(color:reset) %(color:green)%(committerdate:raw)%(color:reset)')
 do
     export IFS=$' '
-    elements=($i)
-    
+    elements=($i)    
     if [ $TIME -gt ${elements[1]} ]
     then
         export IFS="/"
 	    inner_elements=(${elements[0]})
-        git push origin --delete ${inner_elements[1]}
-        #echo ${inner_elements[1]}
-        OUT="${OUT}, ${inner_elements[1]}"
+        if [[ ${inner_elements[1]} != "KEEP"* ]]
+        then
+            git push origin --delete ${inner_elements[1]}
+            # echo ${inner_elements[1]}
+            OUT="${OUT}, ${inner_elements[1]}"
+        fi
     fi
 done
 export IFS=$'\n'
